Add carousel functionality with style
The commit includes creating new files: script.js, style.css, and index.html. In the script file, a carousel functionality was added which includes functions for showing, moving to the next and previous slides, and for starting and stopping the auto slide feature based on events. In the style and html files, designs and structures for the carousel were added respectively.